backend/requirements.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
alembic
pydantic>=2.0
pydantic-settings
python-dotenv
langchain>=0.3
langchain-core>=0.3
langchain-google-genai
google-genai
langsmith
langgraph>=0.4
tavily-python
crawl4ai
pdfplumber
pypdfium2
pgvector
geoalchemy2
geopy
httpx
tenacity
aiofiles
python-multipart
